+++ sources/framework/request/common/add-map-layer-request.js
@@ -14,22 +14,12 @@ Oskari.clazz.define('Oskari.mapframework.request.common.AddMapLayerRequest',
      *
      * @param {String}
      *            mapLayerId id of the map layer (matching one in Oskari.mapframework.service.MapLayerService)
-     * @param {Boolean}
-     *            keepLayersOrder should order of layers be reserved (optional,
-     * defaults to false)
-     * @param {Boolean}
-     *            isBasemap (optional, defaults to false)
-     * @param {Boolean}
-     *            isExternal (optional, not used in paikkatietoikkuna)
      */
 
-    function (mapLayerId, keepLayersOrder, isBasemap, isExternal) {
+    function (mapLayerId) {
 
         this._creator = null;
         this._mapLayerId = mapLayerId;
-        this._keepLayersOrder = (keepLayersOrder == true);
-        this._isExternal = (isExternal == true);
-        this._isBasemap = (isBasemap == true);
     }, {
         /** @static @property __name request name */
         __name: "AddMapLayerRequest",
@@ -47,28 +37,6 @@ Oskari.clazz.define('Oskari.mapframework.request.common.AddMapLayerRequest',
          */
         getMapLayerId: function () {
             return this._mapLayerId;
-        },
-        /**
-         * @method getKeepLayersOrder
-         * @return {Boolean} boolean true if we should keep the layer order
-         */
-        getKeepLayersOrder: function () {
-            return this._keepLayersOrder;
-        },
-        /**
-         * @method isBasemap
-         * @return {Boolean} boolean true if this is a basemap
-         */
-        isBasemap: function () {
-            return this._isBasemap;
-        },
-        /**
-         * @method isExternal
-         * @return {Boolean} true if this is an externally added layer (not found in
-         * MapLayerService?)
-         */
-        isExternal: function () {
-            return this._isExternal;
         }
     }, {
         /**